complete spark2.3 keyed reader
