+++ spark/v2.3/spark/src/main/java/com/netease/arctic/spark/util/ArcticSparkUtil.java
@@ -1,3 +1,21 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package com.netease.arctic.spark.util;
 
 import org.apache.avro.generic.GenericData;
@@ -6,6 +24,7 @@ import org.apache.commons.lang.StringUtils;
 import org.apache.iceberg.types.Type;
 import org.apache.iceberg.util.ByteBuffers;
 import org.apache.spark.sql.RuntimeConfig;
+import org.apache.spark.sql.types.BinaryType;
 import org.apache.spark.sql.types.Decimal;
 import org.apache.spark.unsafe.types.UTF8String;
 
@@ -56,4 +75,19 @@ public class ArcticSparkUtil {
     }
     return value;
   }
+
+  public static Object[] convertRowObject(Object[] objects) {
+    for (int i = 0; i < objects.length; i++) {
+      Object object = objects[i];
+      if (object instanceof UTF8String) {
+        objects[i] = object.toString();
+      } else if (object instanceof Decimal) {
+        objects[i] = ((Decimal) object).toJavaBigDecimal();
+      } else if (object instanceof BinaryType) {
+        objects[i] = ByteBuffer.wrap((byte[]) object);
+      }
+    }
+    return objects;
+  }
+
 }
